+++ cmd/create/appcr/runner.go
@@ -6,15 +6,19 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/spf13/cobra"
 
-	"github.com/giantswarm/app/pkg/app"
+	"github.com/giantswarm/app/v3/pkg/app"
 )
 
 func runAppCRError(cmd *cobra.Command, args []string) error {
-	name := cmd.Flag("name").Value.String()
-	appName := cmd.Flag("app-name").Value.String()
-	appNamespace := cmd.Flag("app-namespace").Value.String()
-	appVersion := cmd.Flag("app-version").Value.String()
-	catalog := cmd.Flag("catalog").Value.String()
+	var (
+		appName       = cmd.Flag("app-name").Value.String()
+		appNamespace  = cmd.Flag("app-namespace").Value.String()
+		appVersion    = cmd.Flag("app-version").Value.String()
+		catalog       = cmd.Flag("catalog").Value.String()
+		name          = cmd.Flag("name").Value.String()
+		configMapName = cmd.Flag("configmap-name").Value.String()
+		secretName    = cmd.Flag("secret-name").Value.String()
+	)
 
 	disableForceUpgrade, err := cmd.Flags().GetBool("disable-force-upgrade")
 	if err != nil {
@@ -30,6 +34,8 @@ func runAppCRError(cmd *cobra.Command, args []string) error {
 		AppVersion:          appVersion,
 		DisableForceUpgrade: disableForceUpgrade,
 		Name:                name,
+		ConfigMapName:       configMapName,
+		SecretName:          secretName,
 	}
 
 	appCR := app.NewCR(config)