+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow app CR configmap and secret configuration.
+
 ## [3.0.6] - 2020-11-13
 
 ### Added